+++ changelog.md
@@ -13,4 +13,8 @@ Initial release
 - Ported to MC 1.21.3
 - Changed how the mod interacts with Cloth-Gamerules
 ### 1.2.1
-- Updated for MC 1.21.11
+- Marked as incompatible with MC 1.21.11
+
+# v2
+- Ported to MC 1.21.11
+- Config file now list gamerule names as namespaced identifier.